+++ app/src/main/java/com/example/itshalloween/ui/screens/main/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.example.itshalloween
+package com.example.itshalloween.ui.screens.main
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -24,13 +24,12 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.itshalloween.R
 
 val halloweenImages = listOf(
     R.drawable.img_sample,
